requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.29.0
python-multipart==0.0.6

# LangGraph & LangChain
langgraph==0.2.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

# PDF Processing
pypdf2==3.0.1
pdf2image==1.16.3
pymupdf==1.23.8

# Data Handling
pydantic==2.5.2
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Utilities
pillow==10.1.0


